+++ HISTORY.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.6] - 2022-11-25
+
 ### Added
 
 - Support for the new `onTimeout` callback
@@ -62,7 +64,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release.
 
-[unreleased]: https://github.com/Le0Developer/react-turnstile/compare/v1.0.5...HEAD
+[unreleased]: https://github.com/Le0Developer/react-turnstile/compare/v1.0.6...HEAD
+[1.0.6]: https://github.com/le0developer/react-turnstile/compare/v1.0.5...v1.0.6
 [1.0.5]: https://github.com/le0developer/react-turnstile/compare/v1.0.4...v1.0.5
 [1.0.4]: https://github.com/le0developer/react-turnstile/compare/v1.0.3...v1.0.4
 [1.0.3]: https://github.com/le0developer/react-turnstile/compare/v1.0.2...v1.0.3